requirements.txt
# GPU-accelerated inference
torch>=2.1.0
faster-whisper>=1.0.0
transformers>=4.35.0
accelerate>=0.25.0

# Audio processing
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.14
webrtcvad>=2.0.10

# TTS
TTS>=0.21.0  # Coqui TTS - works with Python 3.11

# Computer control
pywinauto>=0.6.8
pyautogui>=0.9.54
pygetwindow>=0.0.9
keyboard>=0.13.5
mouse>=0.7.1

# Web automation
playwright>=1.40.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0
requests>=2.31.0
pillow>=10.1.0

# Database
sqlalchemy>=2.0.23

# Vector DB & RAG
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Development
pytest>=7.4.3
black>=23.12.0

# Backend server
flask>=3.0.0
flask-cors>=4.0.0

# API layer
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Process management
psutil>=5.9.0

